refactor: move help screens to `/ui` directory
