+++ bin/gui/main.qml
@@ -21,7 +21,7 @@ import QtQuick.Controls
 import QtQuick.Layouts
 import Qt.labs.qmlmodels
 import "qrc:/ui/about" as About
-import "qrc:/qt/qml/Pepp/gui/helpview" as Help
+import "qrc:/ui/help" as Help
 import "qrc:/ui/memory/hexdump" as Memory
 import "qrc:/ui/cpu" as Cpu
 import "qrc:/qt/qml/Pepp/gui/project"
