+++ ui/CMakeLists.txt
@@ -1,6 +1,7 @@
 add_subdirectory(utils)
 add_subdirectory(about)
 add_subdirectory(text)
+add_subdirectory(help/)
 add_subdirectory(memory)
 add_subdirectory(cpu)
 add_subdirectory(preferences)
@@ -8,7 +9,7 @@ add_subdirectory(project)
 add_subdirectory(mode)
 
 
-set(ui_targets "ui_utils;ui_about;text;memory;cpu;prefs;project;mode")
+set(ui_targets "ui_utils;ui_about;text;memory;cpu;prefs;project;mode;ui_help")
 
 
 if (APPLE)
