+++ ui/help/help_globals.hpp
@@ -1,5 +1,6 @@
 /*
- * Copyright (c) 2023-2024 J. Stanley Warford, Matthew McRaven
+ * Copyright (c) 2023 J. Stanley Warford, Matthew McRaven
+ *
  * This program is free software: you can redistribute it and/or modify
  * it under the terms of the GNU General Public License as published by
  * the Free Software Foundation, either version 3 of the License, or
@@ -14,12 +15,12 @@
  * along with this program.  If not, see <https://www.gnu.org/licenses/>.
  */
 
-#include "registration.hpp"
-#include "book_item_model.hpp"
+#pragma once
 
-namespace helpview {
-void registerTypes(QQmlApplicationEngine &engine) {
-  // TODO: Missing translations
-  qmlRegisterType<builtins::BookModel>("edu.pepp", 1, 0, "BookModel");
-}
-} // namespace helpview
+#include <QtCore/QtGlobal>
+
+#if defined(HELP_LIBRARY)
+#define HELP_EXPORT Q_DECL_EXPORT
+#else
+#define HELP_EXPORT Q_DECL_IMPORT
+#endif
